+++ src/main/java/atomicedit/backend/blockprovider/BlockProvider.java
@@ -11,7 +11,7 @@ public interface BlockProvider {
     
     public Volume getVolume();
     
-    public short getBlockAt(int x, int y, int z);
+    public int getBlockAt(int x, int y, int z);
     
     public void doForBlock(ActionForBlock action);
